Added code for calling provider specific code
