--- internal/gardener/shoot/hyperscaler/aws/controlplane.go
@@ -1,19 +1,0 @@
-package aws
-
-import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-// This types are copied from https://github.com/gardener/gardener-extensions/blob/master/controllers/provider-aws/pkg/apis/aws/types_controlplane.go
-
-// ControlPlaneConfig contains configuration settings for the control plane.
-type ControlPlaneConfig struct {
-	metav1.TypeMeta
-
-	// CloudControllerManager contains configuration settings for the cloud-controller-manager.
-	CloudControllerManager *CloudControllerManagerConfig `json:"cloudControllerManager,omitempty"`
-}
-
-// CloudControllerManagerConfig contains configuration settings for the cloud-controller-manager.
-type CloudControllerManagerConfig struct {
-	// FeatureGates contains information about enabled feature gates.
-	FeatureGates map[string]bool
-}
