+++ internal/gardener/shoot/hyperscaler/azure/model.go
@@ -85,3 +85,20 @@ type Zone struct {
 	// +optional
 	NatGateway *NatGateway `json:"natGateway,omitempty"`
 }
+
+// This types are copied from https://github.com/gardener/gardener-extension-provider-azure/blob/master/pkg/apis/azure/types_controlplane.go
+
+// ControlPlaneConfig contains configuration settings for the control plane.
+type ControlPlaneConfig struct {
+	metav1.TypeMeta
+
+	// CloudControllerManager contains configuration settings for the cloud-controller-manager.
+	// +optional
+	CloudControllerManager *CloudControllerManagerConfig `json:"cloudControllerManager,omitempty"`
+}
+
+// CloudControllerManagerConfig contains configuration settings for the cloud-controller-manager.
+type CloudControllerManagerConfig struct {
+	// FeatureGates contains information about enabled feature gates.
+	FeatureGates map[string]bool
+}
