+++ internal/gardener/shoot/hyperscaler/gcp/model.go
@@ -61,3 +61,22 @@ type FlowLogs struct {
 	// Metadata configures whether metadata fields should be added to the reported VPC flow logs.
 	Metadata *string `json:"metadata,omitempty"`
 }
+
+// This types are copied from https://github.com/gardener/gardener-extension-provider-gcp/blob/master/pkg/apis/gcp/types_controlplane.go
+
+// ControlPlaneConfig contains configuration settings for the control plane.
+type ControlPlaneConfig struct {
+	metav1.TypeMeta
+
+	// Zones are the GCP zones.
+	Zone string `json:"zone"`
+
+	// CloudControllerManager contains configuration settings for the cloud-controller-manager.
+	CloudControllerManager *CloudControllerManagerConfig `json:"cloudControllerManager,omitempty"`
+}
+
+// CloudControllerManagerConfig contains configuration settings for the cloud-controller-manager.
+type CloudControllerManagerConfig struct {
+	// FeatureGates contains information about enabled feature gates.
+	FeatureGates map[string]bool
+}
