+++ internal/gardener/shoot/hyperscaler/sapconvergedcloud/providerspecificconfig.go
@@ -1,0 +1,11 @@
+package sapconvergedcloud
+
+import imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
+
+func GetInfrastructureConfig(runtimeShoot imv1.RuntimeShoot) ([]byte, error) {
+	return nil, nil
+}
+
+func GetControlPlaneConfig(runtimeShoot imv1.RuntimeShoot) ([]byte, error) {
+	return nil, nil
+}
